Refactor HED module
This commit refactors the HED module completely by splitting it into
smaller functions and creating a new function that checks for HED
data in the sidecars and TSV event files before loading a HED schema,
thus addressing #1205. I've also taken the opportunity to fix #1292
by passing jsonFiles from fullTest.js through to the HED module and
using that to find the file parameter to Issue, thus providing the
issue with the actual sidecar name.