Revamp HED issue codes
This commit winnows the list of HED issue codes to 7, one of which
is temporary pending the inclusion of proper placeholder validation
in hed-validation v3.3.0. The tests have been restructured, with the
code-specific tests being deleted and the remaining tests updated
to focus on their application to the BIDS-HED interface.