Migrate bulk of HED code to hed-validator
This branch will migrate the majority of the HED-specific code to
the hed-validator codebase, thus freeing the BIDS code of the burden.
Only the minimum required to sustain the interface is kept, including
determining whether HED validation is needed, extracting the schema
version info, and interfacing with the other BIDS code. I will trim
down the testcases further before this PR is merged.

lodash was added as a direct dependency to add the union function.
It already existed as an indirect dependency.