FIX: Expect multiple rule matches for events.json
This test is a bit weird; it seems to have been relying on arbitrary
pruning. There's nothing in the rules or code that now (as of #1940)
should prevent multiple rules from being returned.

If we do want a test to verify pruning, we should come up with a new
one.