Throw and error, do not just log an error, if provided schema is not readable
I think current behavior violates the principle of the least surprise: I ran
validation while providing custom schema source, it executed with many warnings
and errors reported to the screen. It is only after scrolling far up, I saw
that actually my specification was wrong and operation resorted to built-in schema.

amend: removed comment "version is ignored when the network cannot be accessed"